# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MoneyMate Personal Finance API
  version: 0.1.0
  description: |
    MVP REST API contract for MoneyMate Personal Finance.

    All API responses use the same envelope shape:
    `{ "success": true, "data": {}, "meta": {}, "request_id": "req_xxx" }`.

    Error responses use:
    `{ "success": false, "error": { "code": "...", "message": "...", "details": [] }, "request_id": "req_xxx" }`.

    Authentication uses short-lived JWT access tokens in the `Authorization: Bearer <token>` header.
    Login also sets an HTTP-only `moneymate_refresh_token` cookie scoped to `/api/v1/auth`.
    Refresh token values are not returned in JSON responses.

    Financial safety: all market prices in the MVP are manual/mock. The API does not claim real-time prices and does not provide buy/sell recommendations.
servers:
  - url: http://localhost:8080
    description: Local backend
...
paths:
...
  /api/v1/imports/upload:
    post:
      tags: [Imports]
      summary: Upload CSV/XLSX and create import preview
      security:
        - bearerAuth: []
      description: Requires admin or user role. Import jobs are scoped to the authenticated user. Stores import job rows but does not write portfolio data until confirm.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV or XLSX file.
      responses:
        "201":
          description: Import preview result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportPreviewEnvelope"
              examples:
                csv:
                  value:
                    success: true
                    data:
                      job_id: 55555555-5555-5555-5555-555555555555
                      source_type: csv
                      original_filename: assets.csv
                      detected_sections: [holdings, orders, asset_summary, cash]
                      summary:
                        total_rows: 5
                        valid_rows: 5
                        invalid_rows: 0
                      rows:
                        - id: 66666666-6666-6666-6666-666666666666
                          section: orders
                          row_number: 5
                          status: valid
                          raw:
                            ticker: SPY
                            type: Buy
                          normalized:
                            ticker: SPY
                            type: buy
                            price: 733.33
                            units: 0.075204887
                            currency: USD
                            fx_rate_to_idr: 17936.5119
                          errors: []
                    request_id: req_123
  /api/v1/imports/jobs:
    get:
      tags: [Imports]
      summary: List import jobs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Import job history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJobListEnvelope"
  /api/v1/imports/jobs/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Imports]
      summary: Get import job detail with rows
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Import job detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJobDetailEnvelope"
  /api/v1/imports/jobs/{id}/confirm:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      tags: [Imports]
      summary: Confirm import job and write valid rows
      security:
        - bearerAuth: []
      description: Requires admin or user role. Writes user-scoped transactions, cash accounts, manual prices, import row statuses, recalculates holdings in the same database transaction, and records the import audit log where applicable. Instrument metadata remains shared during the assets-module transition.
      responses:
        "200":
          description: Import confirm result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportConfirmEnvelope"
              examples:
                confirmed:
                  value:
                    success: true
                    data:
                      job_id: 55555555-5555-5555-5555-555555555555
                      status: partial
                      total_rows: 4
                      imported_rows: 3
                      skipped_rows: 0
                      failed_rows: 1
                      holdings_recalculated: true
                      holdings_snapshot_date: "2026-07-02"
                      holdings_count: 1
                      message: Import selesai. Nilai portfolio sudah dihitung ulang. Data harga tetap manual dan bukan real-time.
                    request_id: req_123
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived JWT access token from `/api/v1/auth/login` or `/api/v1/auth/refresh`.
    refreshCookie:
      type: apiKey
      in: cookie
      name: moneymate_refresh_token
      description: HTTP-only refresh cookie set by `/api/v1/auth/login`, scoped to `/api/v1/auth`.
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success, request_id]
      properties:
        success:
          type: boolean
          const: true
        data:
          description: Endpoint-specific response data.
        meta:
          description: Optional endpoint metadata.
        request_id:
          type: string
...
    ImportJob:
      type: object
      required: [id, source_type, status, total_rows, success_rows, failed_rows, created_at]
      properties:
        id:
          type: string
          format: uuid
        source_type:
          type: string
          enum: [csv, xlsx, google_sheet, manual_seed]
        original_filename:
          type:
            - string
            - "null"
        status:
          type: string
          enum: [pending, processing, completed, failed, partial]
        total_rows:
          type: integer
        success_rows:
          type: integer
        failed_rows:
          type: integer
        error_summary:
          type:
            - string
            - "null"
        created_by:
          type:
            - string
            - "null"
          format: uuid
        created_at:
          type: string
          format: date-time
        completed_at:
          type:
            - string
            - "null"
          format: date-time
    ImportPreviewRow:
      type: object
      required: [section, row_number, status, raw, normalized, errors]
      properties:
        id:
          type: string
          format: uuid
        section:
          type: string
          enum: [holdings, orders, asset_summary, cash]
        row_number:
          type: integer
        status:
          type: string
          enum: [valid, invalid, imported, skipped]
        raw:
          type: object
          additionalProperties:
            type: string
        normalized:
          type: object
          additionalProperties: true
        errors:
          type: array
          items:
            type: string
    ImportPreview:
      type: object
      required: [job_id, source_type, original_filename, detected_sections, rows, summary]
      properties:
        job_id:
          type: string
          format: uuid
        source_type:
          type: string
          enum: [csv, xlsx]
        original_filename:
          type: string
        detected_sections:
          type: array
          items:
            type: string
            enum: [holdings, orders, asset_summary, cash]
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ImportPreviewRow"
        summary:
          type: object
          required: [total_rows, valid_rows, invalid_rows]
          properties:
            total_rows:
              type: integer
            valid_rows:
              type: integer
            invalid_rows:
              type: integer
    ImportPreviewEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ImportPreview"
    ImportJobListEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/ImportJob"
    ImportJobDetail:
      allOf:
        - $ref: "#/components/schemas/ImportJob"
        - type: object
          properties:
            detected_sections:
              type: array
              items:
                type: string
                enum: [holdings, orders, asset_summary, cash]
            rows:
              type: array
              items:
                $ref: "#/components/schemas/ImportPreviewRow"
    ImportJobDetailEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ImportJobDetail"
    ImportConfirmEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              required:
                - job_id
                - status
                - total_rows
                - imported_rows
                - skipped_rows
                - failed_rows
                - holdings_recalculated
                - holdings_snapshot_date
                - holdings_count
                - message
              properties:
                job_id:
                  type: string
                  format: uuid
                status:
                  type: string
                total_rows:
                  type: integer
                imported_rows:
                  type: integer
                skipped_rows:
                  type: integer
                failed_rows:
                  type: integer
                holdings_recalculated:
                  type: boolean
                  description: True when holdings snapshots were recalculated in the same database transaction as import confirmation.
                holdings_snapshot_date:
                  type: string
                  format: date
                holdings_count:
                  type: integer
                message:
                  type: string
...